requirements.txt
beautifulsoup4==4.10.0
yagooglesearch==1.6.0
requests==2.27.1